Replaced getSession with getServerAuthSession
