+++ package.json
@@ -38,6 +38,7 @@
     "lucide-react": "^0.279.0",
     "next": "^13.5.2",
     "next-auth": "^4.23.0",
+    "next-superjson-plugin": "^0.5.9",
     "nodemailer": "^6.9.5",
     "radash": "^11.0.0",
     "react": "^18.2.0",
